Blue tinted the buff window buff labels to better match the blue background of buffs.  And be a bit more visible.
Lopped off the whole right side of the group window with all the personal stats and resists and such.  I like the idea of big phat bars for healing, but if I wanna see all that other I can just hit I, don't need it onscreen all the time.

Fix the pet gauge for the local player